β-reduce even when variable is used multiple times
Avoiding β-reduction when a variable occurred multiple times in the lambda body was a necessary optimization for readability back when the presupposition notation was common, but now with effects we will generally keep arguments separate from their presuppositions and small enough to be reasonable to substitute. Stopping short of performing a β-reduction could cause you to miss out on a bunch of other potential reductions, too.