Only track used variables in the scope field
Now a closed expression can be substituted as-is into λ expressions without any rewriting of its scope, which will enable future improvements to the referential stability of reducePass. And substitute can also avoid walking the entire expression tree and instead navigate to exactly the parts of the expression that use the variable, simply by inspecting scope.